+++ src/widgets/calendar.ts
@@ -2,12 +2,15 @@ import { register, type BaseProps, type Widget } from './widget.js';
 import Gtk from 'gi://Gtk?version=3.0';
 
 type Event<Self> = (self: Self) => void
+type Detail<Self> = (self: Self, year: number, month: number, day: number) => string | null
+
 
 export type CalendarProps<
     Attr = unknown,
     Self = Calendar<Attr>,
 > = BaseProps<Self, Gtk.Calendar.ConstructorProperties & {
     on_day_selected?: Event<Self>
+    detail?: Detail<Self>,
 }, Attr>;
 
 export function newCalendar<
@@ -24,6 +27,7 @@ export class Calendar<Attr> extends Gtk.Calendar {
             properties: {
                 'date': ['jsobject', 'r'],
                 'on-day-selected': ['jsobject', 'rw'],
+                'detail': ['jsobject', 'rw'],
             },
         });
     }
@@ -40,6 +44,12 @@ export class Calendar<Attr> extends Gtk.Calendar {
 
     get on_day_selected() { return this._get('on-day-selected') || (() => false); }
     set on_day_selected(callback: Event<this>) { this._set('on-day-selected', callback); }
+
+    get detail() { return this._get('detail-func'); }
+    set detail(func: Detail<this>) {
+        this._set('detail-func', func);
+        this.set_detail_func((self, ...date) => func(self as this, ...date));
+    }
 }
 
 export default Calendar;