+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # git
 
+## Features
+
+- Calendar.detail
+- SpinButton.range
+- SpinButton.increments
+
 ## Fixes
 
 - compiles with typescript >= 5.0.4
+- DrawingArea.draw-fn
 
 # 1.8.0