+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - compiles with typescript >= 5.0.4
 - DrawingArea.draw-fn
+- hyprland: active client empty on window close
 
 # 1.8.0