+++ CHANGELOG.md
@@ -5,12 +5,15 @@
 - Calendar.detail
 - SpinButton.range
 - SpinButton.increments
+- Network.frequency
+- recursive Utils.monitorFile
 
 ## Fixes
 
 - compiles with typescript >= 5.0.4
 - DrawingArea.draw-fn
 - hyprland: active client empty on window close
+- dispose signal on Variable
 
 # 1.8.0